+++ core-ng/src/main/java/core/framework/api/template/TemplateEngine.java
@@ -6,6 +6,8 @@ import core.framework.api.util.StopWatch;
 import core.framework.impl.template.Template;
 import core.framework.impl.template.TemplateBuilder;
 import core.framework.impl.template.location.ClasspathTemplateLocation;
+import core.framework.impl.template.location.StringTemplateLocation;
+import core.framework.impl.template.location.TemplateLocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,11 +32,19 @@ public final class TemplateEngine {
     }
 
     public void addClasspathTemplate(String name, String templateClasspath, Class<?> modelClass) {
+        add(name, new ClasspathTemplateLocation(templateClasspath), modelClass);
+    }
+
+    public void addStringTemplate(String name, String template, Class<?> modelClass) {
+        add(name, new StringTemplateLocation(template), modelClass);
+    }
+
+    private void add(String name, TemplateLocation location, Class<?> modelClass) {
         StopWatch watch = new StopWatch();
         try {
-            templates.put(name, new TemplateBuilder(new ClasspathTemplateLocation(templateClasspath), modelClass).build());
+            templates.put(name, new TemplateBuilder(location, modelClass).build());
         } finally {
-            logger.info("add classpath template, name={}, modelClass={}, elapsedTime={}", name, modelClass.getCanonicalName(), watch.elapsedTime());
+            logger.info("add template, name={}, modelClass={}, elapsedTime={}", name, modelClass.getCanonicalName(), watch.elapsedTime());
         }
     }
 }