+++ core-ng/src/main/java/core/framework/impl/web/response/TextBodyResponseHandler.java
@@ -1,15 +1,16 @@
 package core.framework.impl.web.response;
 
 import core.framework.api.web.ResponseImpl;
-import io.undertow.server.HttpServerExchange;
+import core.framework.impl.web.RequestImpl;
+import io.undertow.io.Sender;
 
 /**
  * @author neo
  */
 public class TextBodyResponseHandler implements BodyHandler {
     @Override
-    public void handle(ResponseImpl response, HttpServerExchange exchange) {
+    public void handle(ResponseImpl response, Sender sender, RequestImpl request) {
         TextBody body = (TextBody) response.body;
-        exchange.getResponseSender().send(body.text);
+        sender.send(body.text);
     }
 }