+++ core-ng/src/test/java/core/framework/impl/queue/RabbitMQEndpointTest.java
@@ -15,6 +15,14 @@ public class RabbitMQEndpointTest {
     }
 
     @Test
+    public void defaultExchange() {
+        // in rabbitmq default exchange is direct and route message to queueName=routingKey
+        RabbitMQEndpoint destination = new RabbitMQEndpoint("rabbitmq://exchange/");
+        Assert.assertEquals("", destination.exchange);
+        Assert.assertEquals("", destination.routingKey);
+    }
+
+    @Test
     public void queue() {
         RabbitMQEndpoint destination = new RabbitMQEndpoint("rabbitmq://queue/test");
         Assert.assertEquals("", destination.exchange);